fix setup.sh

+++ setup.sh
@@ -12,31 +12,25 @@ if [ "$contest" == "" ]; then
     exit 1
 fi
 
-if [ -d $contest ]; then
-    echo "$contest directory already exists."
-    exit 1
-fi
-
 echo $contest
 
 for i in a b c d e f; do
     dir=$contest/$i
     mkdir -p $dir
-    cp boilerplate/main.cpp $dir/main.cpp
+    [[ ! -e $dir/main.cpp ]] && cp boilerplate/main.cpp $dir/main.cpp
 
-
-    cat > $dir/runtest <<EOS
+    cat > $dir/runtest <<'EOS'
 #!/bin/bash
 
 set -e
 
 runtest() {
-    ./a.out < "\$1" > result
-    if [ "\$(cat result)" != "\$(cat \$2)" ]; then
-        echo "\$1: \$(cat \$2) expected, but got \$(cat result)"
+    ./a.out < "$1" > result
+    if [ "$(cat result)" != "$(cat $2)" ]; then
+        echo "$1: $(cat $2) expected, but got $(cat result)"
         exit 1
     fi
-    echo "test case #\$1 => \$(cat result)"
+    echo "test case #$1 => $(cat result)"
 }
 
 clang++ -Wall -O2 main.cpp
